+++ src/Command/PopulateCommand.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Setono\SyliusAlgoliaPlugin\Command;
 
-use Setono\SyliusAlgoliaPlugin\IndexResolver\ProductIndexesResolverInterface;
 use Setono\SyliusAlgoliaPlugin\Message\Command\PopulateProductIndex;
+use Setono\SyliusAlgoliaPlugin\Provider\ProductIndexScopeProviderInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -17,23 +17,23 @@ final class PopulateCommand extends Command
 
     protected static $defaultDescription = 'Populate indexes';
 
-    private ProductIndexesResolverInterface $productIndexesResolver;
+    private ProductIndexScopeProviderInterface $productIndexScopeProvider;
 
     private MessageBusInterface $commandBus;
 
     public function __construct(
-        ProductIndexesResolverInterface $productsIndicesResolver,
+        ProductIndexScopeProviderInterface $productIndexScopeProvider,
         MessageBusInterface $messageBus
     ) {
         parent::__construct();
 
-        $this->productIndexesResolver = $productsIndicesResolver;
+        $this->productIndexScopeProvider = $productIndexScopeProvider;
         $this->commandBus = $messageBus;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $productIndexScopes = $this->productIndexesResolver->resolve();
+        $productIndexScopes = $this->productIndexScopeProvider->resolve();
 
         foreach ($productIndexScopes as $productIndexScope) {
             $this->commandBus->dispatch(new PopulateProductIndex($productIndexScope));
