+++ src/Resources/config/services.xml
@@ -10,6 +10,7 @@
         <import resource="services/event_subscriber.xml"/>
         <import resource="services/index_resolver.xml"/>
         <import resource="services/message.xml"/>
+        <import resource="services/provider.xml"/>
         <import resource="services/settings_provider.xml"/>
     </imports>
 </container>
