+++ src/Resources/config/services/provider.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<container xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://symfony.com/schema/dic/services"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="setono_sylius_algolia.provider.product_index_scope"
+                 class="Setono\SyliusAlgoliaPlugin\Provider\ProductIndexScopeProvider">
+            <argument type="service" id="sylius.repository.channel"/>
+        </service>
+        <service id="Setono\SyliusAlgoliaPlugin\Provider\ProductIndexScopeProviderInterface"
+                 alias="setono_sylius_algolia.provider.product_index_scope"/>
+    </services>
+</container>
